requirements.txt
SQLAlchemy
requests
beautifulsoup4
pymysql
cryptography
sytk
pdfminer.six